lint and move test class
